Re-organize logos & improved styling
