+++ client/src/components/Input/Input.css
@@ -23,7 +23,7 @@
   z-index: 10;
   position: fixed;
   right: 4.2vw;
-  bottom: 21px;
-  height: 35px;
+  bottom: 23px;
+  height: 30px;
   margin: 0;
 }
